pyproject.toml
[tool.poetry]
name = "wuneed"
version = "0.1.0"
description = "A powerful and modular CLI tool for developers"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.7.0"
rich = "^13.0.0"
pydantic = "^1.10.0"
PyYAML = "^6.0"
pytube = "^12.1.0"
whisper = "^1.1.10"
openai = "^0.27.0"
sentence-transformers = "^2.2.2"
weaviate-client = "^3.15.0"
textual = "^0.9.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
wuneed = "wuneed.cli.main:app"